Use centralised configuration; fix bugs in EMPI; namespaces!
Change to using aero library for configuration - push secrets out to a .secrets.edn file in home directory. Create a centralised repository of identifier namespace lookup - not mandatory to use but permits validation and acts as documentation.
Fixed: Generate UUID and date/time for each EMPI request, not just when application started!